requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
commandline-config==2.2.3
Faker==17.6.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
prettytable==3.6.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
pytest==7.2.1
pytest-repeat==0.9.1
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
webdriver-manager==3.8.5
wsproto==1.2.0
